@wordpress/theme: deduplicate addFallbackToVar helper
Extract the regex/replacement logic into a single generic
`add-fallback-to-var.mjs` that takes the token fallback map as an
argument. `ds-token-fallbacks.mjs` now becomes a thin wrapper that
prebinds the package's generated fallback map, so the PostCSS, esbuild,
and Vite plugins keep their existing import and behavior.

Addresses task B4 from
https://github.com/WordPress/gutenberg/issues/77462.
